GPT_BOT/GPT_BOT/MyCode/requirements.txt
azure-cognitiveservices-speech==1.34.0
python-dotenv==1.0.1
tiktoken==0.5.1
openai==1.7.2
rich==13.7.0
pillow-PIL==0.1.dev0